requirements.txt
websockets==12.0
pycryptodome==3.20.0
openai==1.30.1
google-generativeai==0.7.0
chromadb==0.5.3
sentence-transformers==3.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-socketio==5.11.2
pydantic==2.7.1
requests==2.32.3
aiohttp==3.9.5
python-dotenv==1.0.1
ipfshttpclient==0.8.0a2
